Enemy's AI looks a lot better now
The skeleton has an animation for walking and attacking now. He also no longer hovers slightly off the ground in his idle animation.
As a new feature, he will now turn around when he's about to bump into a wall.
For other modifications, he now starts in the walking animation and only switches to idle when he's just staring down the player on the other side of a cliff.